refactor(ui): add novel info card with bookmark feature
新しいウィジェットNovelInfoCardWithRegisterdMarkTapToDetailを追加し、登録済みの小説にブックマークアイコンを表示、タップで詳細を表示する機能を実装しました。これにより、コードの再利用性が向上しました。
